backend/requirements.txt
fastapi
uvicorn[standard]
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
anthropic
python-dotenv
